style: change style

+++ src/main/java/com/oneday/core/service/PaymentService.java
@@ -2,6 +2,7 @@ package com.oneday.core.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 import lombok.RequiredArgsConstructor;
 
 import java.time.OffsetDateTime;
@@ -9,6 +10,7 @@ import java.time.format.DateTimeFormatter;
 import java.util.Map;
 
 import com.oneday.core.entity.Payment;
+import com.oneday.core.entity.Reservation;
 import com.oneday.core.repository.PaymentRepository;
 
 @Service
@@ -23,22 +25,22 @@ public class PaymentService {
 	 * 결제 승인 후, 예약 생성과 결제 정보 저장을 동시 처리
 	 */
 	@Transactional
-	public Payment createReservationAndPayment(int classId, int studentId, Map<String, Object> tossResponse) {
+	public Payment createReservationAndPayment(int classId, long studentId, Map<String, Object> tossResponse) {
 
 		Reservation savedReservation = reservationService.createReservation(classId, studentId);
 		// 결제 정보 저장
 		DateTimeFormatter formatter = DateTimeFormatter.ISO_OFFSET_DATE_TIME;
 
 		Payment newPayment = Payment.builder()
-			.reservation(savedReservation)
-			.tossOrderId((String) tossResponse.get("orderId"))
-			.tossPaymentKey((String) tossResponse.get("paymentKey"))
-			.tossPaymentMethod((String) tossResponse.get("method"))
-			.tossPaymentStatus((String) tossResponse.get("status"))
-			.totalAmount(((Number) tossResponse.get("totalAmount")).intValue())
-			.requestedAt(OffsetDateTime.parse((String) tossResponse.get("requestedAt"), formatter).toLocalDateTime())
-			.approvedAt(OffsetDateTime.parse((String) tossResponse.get("approvedAt"), formatter).toLocalDateTime())
-			.build();
+				.reservation(savedReservation)
+				.tossOrderId((String)tossResponse.get("orderId"))
+				.tossPaymentKey((String)tossResponse.get("paymentKey"))
+				.tossPaymentMethod((String)tossResponse.get("method"))
+				.tossPaymentStatus((String)tossResponse.get("status"))
+				.totalAmount(((Number)tossResponse.get("totalAmount")).intValue())
+				.requestedAt(OffsetDateTime.parse((String)tossResponse.get("requestedAt"), formatter).toLocalDateTime())
+				.approvedAt(OffsetDateTime.parse((String)tossResponse.get("approvedAt"), formatter).toLocalDateTime())
+				.build();
 
 		return paymentRepository.save(newPayment);
 	}
